reducer 추가 및 reducer test 코드 작성
